ON-670: Complete Refactor of ERC-7432
