+++ test/SftRolesRegistry/types.ts
@@ -13,25 +13,3 @@ export interface GrantRoleData {
   revocable: boolean
   data: string
 }
-
-export interface RoleAssignment {
-  itemId: number
-  role: string
-  tokenAddress: string
-  tokenId: number
-  tokenAmount: number
-  grantor: string
-  grantee: string
-  expirationDate: number
-  revocable: boolean
-  data: string
-}
-
-export interface RevokeRoleData {
-  itemId: number
-  role: string
-  tokenAddress: string
-  tokenId: number
-  grantor: string
-  grantee: string
-}
