+++ test/types.ts
@@ -1,28 +1,9 @@
-export interface NftMetadata {
-  name: string
-  description: string
-  roles: Role[]
-}
-
 export interface Role {
-  name: string
-  description: string
-  isUniqueRole: boolean
-  inputs: Input[]
-}
-
-export interface Input {
-  name: string
-  type: string
-  components?: Input[]
-}
-
-export interface RoleAssignment {
-  role: string
   tokenAddress: string
   tokenId: number
-  grantor: string
+  roleId: string
   grantee: string
   expirationDate: number
+  revocable: boolean
   data: string
 }
